# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de la Protectora",
        "description": "Documentación de la API para gestión de mascotas",
        "contact": {
            "email": "isa@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/mascotas": {
            "get": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Obtener todas las mascotas",
                "operationId": "b06bcf95c18bee8ed15647d8fa088cbf",
                "responses": {
                    "200": {
                        "description": "Lista de mascotas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Crear una nueva mascota",
                "operationId": "7ef45de4bf722d348a63d7f462b44898",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "especie",
                                    "raza",
                                    "edad",
                                    "descripcion",
                                    "estado"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "especie": {
                                        "type": "string"
                                    },
                                    "raza": {
                                        "type": "string"
                                    },
                                    "edad": {
                                        "type": "integer"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "en proceso de adopción",
                                            "adoptado"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mascota creada exitosamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/mascotas/paginadas": {
            "get": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Obtener mascotas paginadas",
                "operationId": "628134b3612d5f14aef839aa44701e3c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de mascotas por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mascotas paginada"
                    }
                }
            }
        },
        "/api/mascotas/{id}": {
            "get": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Obtener una mascota por ID",
                "operationId": "1a2eaf8302ca37018e3c8449c1b7c9ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mascota encontrada"
                    },
                    "404": {
                        "description": "Mascota no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Actualizar una mascota",
                "operationId": "58258f41d5e256a91ee928955dbd1ad8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "especie": {
                                        "type": "string"
                                    },
                                    "raza": {
                                        "type": "string"
                                    },
                                    "edad": {
                                        "type": "integer"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "en proceso de adopción",
                                            "adoptado"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mascota actualizada"
                    },
                    "404": {
                        "description": "Mascota no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Eliminar una mascota",
                "operationId": "ff434682a0ddbfa68e8fc15a98d51de7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mascota eliminada"
                    },
                    "404": {
                        "description": "Mascota no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}